+++ include/seqan3/io/alignment_file/format_bam.hpp
@@ -124,6 +124,7 @@ public:
               typename ref_id_type,
               typename ref_offset_type,
               typename align_type,
+              typename cigar_type,
               typename flag_type,
               typename mapq_type,
               typename qual_type,
@@ -143,6 +144,7 @@ public:
               ref_id_type                                             & ref_id,
               ref_offset_type                                         & ref_offset,
               align_type                                              & align,
+              cigar_type                                              & cigar_vector,
               flag_type                                               & flag,
               mapq_type                                               & mapq,
               mate_type                                               & mate,
@@ -167,7 +169,7 @@ public:
         // these variables need to be stored to compute the ALIGNMENT
         [[maybe_unused]] int32_t offset_tmp{};
         [[maybe_unused]] int32_t soft_clipping_end{};
-        [[maybe_unused]] std::vector<cigar> cigar_vector{};
+        [[maybe_unused]] std::vector<cigar> tmp_cigar_vector{};
         [[maybe_unused]] int32_t ref_length{0}, seq_length{0}; // length of aligned part for ref and query
 
         // Header
@@ -266,10 +268,11 @@ public:
 
         // read cigar string
         // -------------------------------------------------------------------------------------------------------------
-        if constexpr (!detail::decays_to_ignore_v<align_type>)
+        if constexpr (!detail::decays_to_ignore_v<align_type> || !detail::decays_to_ignore_v<cigar_type>)
         {
-            std::tie(cigar_vector, ref_length, seq_length) = parse_cigar(stream_view, core.n_cigar_op);
-            transfer_soft_clipping_to(cigar_vector, offset_tmp, soft_clipping_end);
+            std::tie(tmp_cigar_vector, ref_length, seq_length) = parse_cigar(stream_view, core.n_cigar_op);
+            transfer_soft_clipping_to(tmp_cigar_vector, offset_tmp, soft_clipping_end);
+            // the actual cigar_vector is swapped with tmp_cigar_vector at the end to avoid copying
         }
         else
         {
@@ -298,7 +301,7 @@ public:
                                   "If you want to read ALIGNMENT but not SEQ, the alignment"
                                   " object must store a sequence container at the second (query) position.");
 
-                    if (!cigar_vector.empty()) // only parse alignment if cigar information was given
+                    if (!tmp_cigar_vector.empty()) // only parse alignment if cigar information was given
                     {
                         assert(core.l_seq == (seq_length + offset_tmp + soft_clipping_end)); // sanity check
                         using alph_t = value_type_t<decltype(get<1>(align))>;
@@ -413,9 +416,9 @@ public:
                                        "record.")};
 
                     auto cigar_view = std::views::all(std::get<std::string>(it->second));
-                    std::tie(cigar_vector, ref_length, seq_length) = sam_fmt::parse_cigar(cigar_view);
-                    offset_tmp = 0, soft_clipping_end = 0;
-                    transfer_soft_clipping_to(cigar_vector, offset_tmp, soft_clipping_end);
+                    std::tie(tmp_cigar_vector, ref_length, seq_length) = sam_fmt::parse_cigar(cigar_view);
+                    offset_tmp = soft_clipping_end = 0;
+                    transfer_soft_clipping_to(tmp_cigar_vector, offset_tmp, soft_clipping_end);
 
                     assign_unaligned(get<1>(align),
                                      seq | views::slice(static_cast<decltype(std::ranges::distance(seq))>(offset_tmp),
@@ -425,8 +428,11 @@ public:
             }
 
             // Alignment object construction
-            construct_alignment(align, cigar_vector, core.refID, ref_seqs, core.pos, ref_length); // inherited from SAM format
+            construct_alignment(align, tmp_cigar_vector, core.refID, ref_seqs, core.pos, ref_length); // inherited from SAM format
         }
+
+        if constexpr (!detail::decays_to_ignore_v<cigar_type>)
+            std::swap(cigar_vector, tmp_cigar_vector);
     }
 
 protected: