+++ CHANGELOG.md
@@ -31,6 +31,7 @@ If possible, provide tooling that performs the changes, e.g. a shell-script.
 #### Input/Output
 
 * Asynchronous input (background file reading) supported via seqan3::view::async_input_buffer.
+* Reading field::CIGAR into a vector over seqan3::cigar is supported via seqan3::alignment_file_input.
 
 ## API changes